+++ belokurskaya.sofia/T1/rectangle.cpp
@@ -2,8 +2,7 @@
 
 #include <stdexcept>
 
-Rectangle::Rectangle(const point_t & pos, float width, float height):
-pos_(pos), width_(width), height_(height)
+Rectangle::Rectangle(const point_t & pos, double width, double height):pos_(pos), width_(width), height_(height)
 {
   if (width <= 0 || height <= 0)
   {
@@ -11,7 +10,7 @@ pos_(pos), width_(width), height_(height)
   }
 }
 
-float Rectangle::getArea() const
+double Rectangle::getArea() const
 {
   return width_ * height_;
 }
@@ -26,13 +25,13 @@ void Rectangle::move(const point_t & new_pos)
   pos_ = new_pos;
 }
 
-void Rectangle::move(float dx, float dy)
+void Rectangle::move(double dx, double dy)
 {
   pos_.x += dx;
   pos_.y += dy;
 }
 
-void Rectangle::scale(float factor)
+void Rectangle::scale(double factor)
 {
   if (factor <= 0)
   {
