+++ belokurskaya.sofia/T1/shape.hpp
@@ -8,15 +8,15 @@ class Shape
   public:
     virtual ~Shape() = default;
 
-    virtual float getArea() const = 0;
+    virtual double getArea() const = 0;
 
     virtual rectangle_t getFrameRect() const = 0;
 
     virtual void move(const point_t & new_pos) = 0;
 
-    virtual void move(float dx, float dy) = 0;
+    virtual void move(double dx, double dy) = 0;
 
-    virtual void scale(float factor) = 0;
+    virtual void scale(double factor) = 0;
 };
 
 #endif
